\name{B2M}
\alias{B2M}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert CpG Beta value to M value
}
\description{
Convert CpG Beta value to M value
}
\usage{
B2M(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Matrix of CpG Beta values 
}
}

\value{
Matrix of CpG M values
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
CpGFilter:Model-based CpG probe filtering with replicates for epigenome-wide association studies (2014). Submitted.
}
